Changed to use a 3D array of floats, instead of ints, so that the data can be passed in efficiently, without needing to make a copy of the data as an array of floats. Changed DOS to UNIX text format.

+++ trunk/Operators/TOF_SCD/INTGT_SLICE.java
@@ -5,7 +5,7 @@ import Command.JavaCC.Fortran.*;
 
 public class INTGT_SLICE implements Wrappable,HiddenOperator{
 
-  public int[][][] JHIST=new int[0][0][0];
+  public float[][][] JHIST=new float[0][0][0];
 
   public int[][] RANGE=new int[0][0];
